Added minHeight when content have zero height
