+++ core-web/libs/sdk/react/src/lib/components/Container/Container.module.css
@@ -5,48 +5,3 @@
  * .container {
  * }
 */
-
-/* div[data-dot-object='']{
-    container: parent / size;
-} */
-
-/* @container  (height < 1px) {
-    div[data-dot-object='contentlet'] {
-      border: 1px solid red;
-      min-height: 32px;
-      min-width: 500px;
-    }
-  } */
-/* div[data-dot-object='contentlet']:empty {
-    min-height: 32px;
-    min-width: 100%;
-}
-
-div[data-dot-object='contentlet']:empty::after {
-    content: 'Contentlet component with empty HTML content';
-    color: rgb(3, 14, 50);
-    display: block;
-    padding: 10px;
-    text-align: center;
-    background-color: rgb(236, 240, 253);
-} */
-
-
-div[data-dot-object='contentlet'] > * {
-    container: parent / size;
-    /* height: 100%; */
-}
-
-.empty-contentlet {
-    min-height: 32px;
-    min-width: 100%;
-}
-
-
-@container (height < 1px) {
-    div[data-dot-object='contentlet'] {
-      border: 1px solid red;
-      min-height: 32px;
-      min-width: 500px;
-    }
-  }
